+++ src/main/kotlin/com/babytrackr/analytics/infrastructure/model/FeedPayload.kt
@@ -1,5 +1,5 @@
 package com.babytrackr.analytics.infrastructure.model
 
 data class FeedPayload(
-    val feedingAmount: Int
+    val feedingAmount: Double
 )
